fewer changes on merge
