+++ six/modules/c++/six.sicd/six.sicd.vcxproj.filters
@@ -138,9 +138,6 @@
     <ClInclude Include="include\six\sicd\ComplexToAMP8IPHS8I_KDTree.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="include\six\sicd\AMP8I_PHS8I_Converter.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="pch.cpp">
@@ -252,9 +249,6 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="source\ComplexToAMP8IPHS8I.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="source\AMP8I_PHS8I_Converter.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
